requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
boto3==1.35.0
minio==7.2.0
moviepy==1.0.3
pydantic==2.10.6
python-dotenv==1.0.0
requests==2.32.3
python-multipart==0.0.12
aiohttp==3.10.11
aiofiles==24.1.0

# YouTube API dependencies
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.153.0

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1